requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
gunicorn==21.2.0
google-generativeai>=0.3.0
